Move `AccountStateDelta` to `Lib9c.Formatters`
